move appointment components into features folder
