+++ src/client/features/appointments/hooks/useAppointment.ts
@@ -1,8 +1,8 @@
 import { useCallback } from 'react';
-import  useGlobalContext from './useGlobalContext';
-import { appointmentActions } from '../context/reducers/appointmentReducer';
-import { appointmentServices, useAppointmentService } from '../services/appointmentServices';
-import type { AppointmentData, DateRangeObject } from '../context/types/state';
+import  useGlobalContext from '@client/hooks/useGlobalContext';
+import { appointmentActions } from '@client/context/reducers/appointmentReducer';
+import { appointmentServices, useAppointmentService } from '../appointmentServices';
+import type { AppointmentData, DateRangeObject } from '@client/context/types/state';
 
 export const useAppointment = () => {
   const { state, dispatch } = useGlobalContext();
